chore: remove func-tastic from knative GH actions (#235)
Some of the actions are OK, but others such as `go test` are causing
pull requests to fail. This repo is not a typical golang repo in the
traditional knative sense. It is a set of templates for the `func`
binary and as such does not need most of the regular workflow jobs.

Signed-off-by: Lance Ball <lball@redhat.com>

Signed-off-by: Lance Ball <lball@redhat.com>